+++ nonsoolmateClient/src/socialLogin/components/Interceptors.tsx
@@ -1,0 +1,7 @@
+import { Outlet } from "react-router-dom";
+import useSetInterceptors from "socialLogin/hooks/useSetInterceptors";
+
+export default function Interceptors() {
+  useSetInterceptors();
+  return <Outlet />;
+}
